tests(e2e): bleve indexer ignores all-numeric search terms (#37221)
Rename to repo-search.test.ts per reviewer feedback.
Parallelize login and apiCreateFile with Promise.all.

Ref #37221
